size_t was replaced with uint64_t.

+++ inference-engine/src/transformations/src/transformations/common_optimizations/split_concat_pair_to_interpolate_fusion.cpp
@@ -19,15 +19,15 @@
 
 namespace {
 // This function creates a partition of its argument into groups consisting of adjacent identical elements.
-// Argument: std::vector<size_t> v
+// Argument: std::vector<uint64_t> v
 // Returns: partition of the argument
-std::vector<std::vector<size_t>> grouped_vector(const std::vector<size_t>& v) {
-    std::vector<std::vector<size_t>> result;
+std::vector<std::vector<uint64_t>> grouped_vector(const std::vector<uint64_t>& v) {
+    std::vector<std::vector<uint64_t>> result;
 
     if (v.empty()) return result;
 
-    size_t prev = v[0];
-    std::vector<size_t> group;
+    uint64_t prev = v[0];
+    std::vector<uint64_t> group;
 
     for (const auto& x : v) {
         if (prev != x) {
@@ -41,7 +41,7 @@ std::vector<std::vector<size_t>> grouped_vector(const std::vector<size_t>& v) {
     return result;
 }
 
-std::pair<std::shared_ptr<ngraph::opset8::Split>, int64_t> get_split_before_concat(const std::shared_ptr<ngraph::opset8::Concat>& concat) {
+std::pair<std::shared_ptr<ngraph::opset8::Split>, uint64_t> get_split_before_concat(const std::shared_ptr<ngraph::opset8::Concat>& concat) {
     // This function gets producers of the 'concat' node, checks that the following conditions are fulfilled:
     // 1) all producers for 'concat' are Split nodes;
     // 2) 'concat' has only one unique producer ('split');
@@ -52,7 +52,7 @@ std::pair<std::shared_ptr<ngraph::opset8::Split>, int64_t> get_split_before_conc
     // and, if all these conditions are fulfilled, returns the above mentioned 'Concat' node. Otherwise, if some of these
     // conditions is false, this functions returns nullptr.
 
-    std::vector<size_t> idx;
+    std::vector<uint64_t> idx;
     std::shared_ptr<ngraph::opset8::Split> split;
     for (const auto& input : concat->input_values()) {
         // If 'concat' has some non-Split producer, then the transformation is not applicable.
@@ -60,7 +60,7 @@ std::pair<std::shared_ptr<ngraph::opset8::Split>, int64_t> get_split_before_conc
         if (!split)
             split = split_op;
         if (!split_op || split != split_op) return {};
-        idx.emplace_back(input.get_index());
+        idx.emplace_back(static_cast<uint64_t>(input.get_index()));
     }
 
     // If 'split' node has more than one consumer, then the transformation is not applicable.
@@ -72,19 +72,19 @@ std::pair<std::shared_ptr<ngraph::opset8::Split>, int64_t> get_split_before_conc
 
     // If numbers of consumer ports are various for various output ports of 'split', then the transformation is not applicable.
     auto grouped_idx = grouped_vector(idx);
-    std::unordered_set<size_t> sizes_of_groups;
+    std::unordered_set<uint64_t> sizes_of_groups;
     for (const auto& group : grouped_idx) {
-        sizes_of_groups.insert(group.size());
+        sizes_of_groups.insert(static_cast<uint64_t>(group.size()));
     }
     if (sizes_of_groups.size() != 1) return {};
-    int64_t size_of_group = static_cast<int64_t>(*(sizes_of_groups.begin()));
+    uint64_t size_of_group = *(sizes_of_groups.begin());
 
     // The transformation is applicable iff output port 0 of 'split' goes to ports [0, ..., m-1] of next node,
     // output port 1 of 'split' goes to ports [m, ..., m + (m-1)] of next node, ..., output port i of 'split'
     // goes to ports [i * m, ..., i * m + (m - 1)], and so on.
-    for (size_t i = 0; i < grouped_idx.size(); ++i) {
+    for (uint64_t i = 0; i < static_cast<uint64_t>(grouped_idx.size()); ++i) {
         const auto& current_group = grouped_idx[i];
-        if (std::any_of(current_group.begin(), current_group.end(), [i](size_t j){ return j != i; })) { return {}; }
+        if (std::any_of(current_group.begin(), current_group.end(), [i](uint64_t j){ return j != i; })) { return {}; }
     }
 
     return {split, size_of_group};
@@ -103,7 +103,7 @@ ngraph::pass::SplitConcatPairToInterpolateFusion::SplitConcatPairToInterpolateFu
         auto concat = std::dynamic_pointer_cast<opset8::Concat>(m.get_match_root());
         if (!concat) return false;
 
-        int64_t scale_factor;
+        uint64_t scale_factor;
         std::shared_ptr<opset8::Split> split;
         std::tie(split, scale_factor) = get_split_before_concat(concat);
         if (!split || !scale_factor) return false;
